requirements.txt
appdirs==1.4.4
asgiref==3.3.4
backcall==0.2.0
certifi==2020.12.5
decorator==5.0.7
distlib==0.3.1
Django==3.2
django-extensions==3.1.3
djangorestframework==3.12.4
filelock==3.0.12
flake8==3.9.1
ipython==7.23.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
mccabe==0.6.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2020.11.15
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.8.1
python-decouple==3.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
virtualenv==20.4.4
virtualenv-clone==0.5.4
wcwidth==0.2.5
